+++ CHANGELOG.md
@@ -5,6 +5,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- SPUserProfileServiceApp
+  - Changed MySiteHostLocation parameter mandatory=$false
+  - Added validation to Set function for testing if SiteNamingConflictResolution parameter is defined then
+    also MySiteHostLocation parameters has to be because it is a mandatory parameter in the parameter set of New-SPProfileServiceApplication when SiteNamingConflictResolution is used
+  - Added "MySiteHostLocation" to Test-SPDscParameterState function in Test-TargetResource
+
+
 ### Fixed
 
 - SPWebAppPolicy